Started Factorizing the Preview Shader Options

+++ Runtime/Nodes/MixtureNode.cs
@@ -41,7 +41,7 @@ namespace Mixture
         [SerializeField]
         public float previewMip = 0.0f;
         [SerializeField]
-        public float previewSlice = 0.0f;
+        public int previewSlice = 0;
         [SerializeField]
         public bool previewVisible = true;
